+++ packages/access-definitions/src/Wrapping/PropertyBehavior/Attribute/PathAttributeSetBehavior.php
@@ -7,6 +7,7 @@ namespace EDT\Wrapping\PropertyBehavior\Attribute;
 use EDT\JsonApi\ApiDocumentation\OptionalField;
 use EDT\Querying\Contracts\PathsBasedInterface;
 use EDT\Querying\Contracts\PropertyAccessorInterface;
+use EDT\Wrapping\PropertyBehavior\Attribute\Factory\PathAttributeSetBehaviorFactory;
 use Webmozart\Assert\Assert;
 
 /**
@@ -34,6 +35,21 @@ class PathAttributeSetBehavior extends AbstractAttributeSetBehavior
         OptionalField $optional
     ) {
         parent::__construct($propertyName, $entityConditions, $optional);
+    }
+
+    /**
+     * @template TCond of PathsBasedInterface
+     *
+     * @param list<TCond> $entityConditions
+     *
+     * @return PathAttributeSetBehaviorFactory<TCond>
+     */
+    public static function createFactory(
+        PropertyAccessorInterface $propertyAccessor,
+        array $entityConditions,
+        OptionalField $optional
+    ): PathAttributeSetBehaviorFactory {
+        return new PathAttributeSetBehaviorFactory($propertyAccessor, $entityConditions, $optional);
     }
 
     protected function updateAttributeValue(object $entity, mixed $value): array