+++ packages/access-definitions/src/Wrapping/PropertyBehavior/Relationship/RelationshipSetBehaviorFactoryInterface.php
@@ -16,8 +16,6 @@ use EDT\Querying\Contracts\PathsBasedInterface;
 interface RelationshipSetBehaviorFactoryInterface
 {
     /**
-     * TODO: refactor to __invoke and merge implementations into the corresponding behavior classes
-     *
      * @param non-empty-string $name
      * @param non-empty-list<non-empty-string> $propertyPath
      * @param class-string<TEntity> $entityClass
@@ -25,5 +23,5 @@ interface RelationshipSetBehaviorFactoryInterface
      *
      * @return RelationshipSetBehaviorInterface<TCondition, TSorting, TEntity, TRelationship>
      */
-    public function createRelationshipSetBehavior(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface;
+    public function __invoke(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface;
 }