+++ packages/access-definitions/src/Wrapping/PropertyBehavior/Relationship/ToOne/Factory/CallbackToOneRelationshipSetBehaviorFactory.php
@@ -39,7 +39,7 @@ class CallbackToOneRelationshipSetBehaviorFactory implements RelationshipSetBeha
         $this->setBehaviorCallback = $setBehaviorCallback;
     }
 
-    public function createRelationshipSetBehavior(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface
+    public function __invoke(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface
     {
         return new CallbackToOneRelationshipSetBehavior(
             $name,