+++ packages/access-definitions/src/Wrapping/PropertyBehavior/Relationship/ToOne/Factory/PathToOneRelationshipSetBehaviorFactory.php
@@ -33,7 +33,7 @@ class PathToOneRelationshipSetBehaviorFactory implements RelationshipSetBehavior
         protected readonly array $entityConditions
     ) {}
 
-    public function createRelationshipSetBehavior(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface
+    public function __invoke(string $name, array $propertyPath, string $entityClass, ResourceTypeInterface $relationshipType): RelationshipSetBehaviorInterface
     {
         return new PathToOneRelationshipSetBehavior(
             $name,