+++ packages/jsonapi/src/PropertyConfig/Builder/AbstractPropertyConfigBuilder.php
@@ -23,7 +23,7 @@ use Webmozart\Assert\Assert;
  * @template TEntity of object
  * @template TCondition of PathsBasedInterface
  * @template TValue
- * @template TConstructorBehaviorFactory
+ * @template TConstructorBehaviorFactory of object
  * @template TPostConstructorBehaviorFactory of object
  * @template TUpdateBehaviorFactory of object
  *